Now reporting which of the Nature paper's genes assc. w/significant GOs are also related to cell cycle.

+++ tests/test_genes_cell_cycle.py
@@ -11,9 +11,14 @@ from goatools.wr_tbl import prt_txt
 __copyright__ = "Copyright (C) 2010-2016, DV Klopfenstein, H Tang, All rights reserved."
 __author__ = "DV Klopfenstein"
 
-def test_cell_cycle(log=sys.stdout):
+def test_cell_cycle(taxid=9606, log=sys.stdout):
+    """Get all genes related to cell cycle. Write results to file."""
+    geneids = get_genes_cell_cycle(taxid, log)
+    fout = "cell_cycle_genes_{TAXID}.txt".format(TAXID=taxid)
+    prt_genes(fout, geneids, taxid, log)
+
+def get_genes_cell_cycle(taxid=9606, log=sys.stdout):
     """Test GOEA with local multipletest correction methods."""
-    taxid = 9606 # Human annotations
     # Download ontologies and annotations, if necessary
     fin_go_obo = "go-basic.obo"
     if not os.path.exists(fin_go_obo):
@@ -23,12 +28,13 @@ def test_cell_cycle(log=sys.stdout):
     # taxid2asscs contains both GO2GeneIDs and GeneID2GOs.
     taxid2asscs = defaultdict(lambda: defaultdict(lambda: defaultdict(set)))
     get_assoc_ncbi_taxids([taxid], taxid2asscs=taxid2asscs)
+
     # Initialize GO-search helper object with obo and annotations(go2items)
     srch = GoSearch(fin_go_obo, go2items=taxid2asscs[taxid]['GO2GeneIDs'])
     # Compile search pattern for 'cell cycle'
     cell_cycle = re.compile(r'cell cycle', flags=re.IGNORECASE)
     # Find ALL GOs that have 'cell cycle'. Store results in file.
-    fout_allgos = "cell_cycle_gos.log" # Log the search results
+    fout_allgos = "cell_cycle_gos_{TAXID}.log".format(TAXID=taxid) 
     with open(fout_allgos, "w") as prt:
         # Search for 'cell cycle' in GO terms
         gos_cc_all = srch.get_matching_gos(cell_cycle, prt=prt)
@@ -42,23 +48,38 @@ def test_cell_cycle(log=sys.stdout):
         gos = gos_cc_all.difference(gos_no_cc)
         # Add children GOs of cell cycle GOs
         gos_all = srch.add_children_gos(gos)
-        log.write('    FOUND {N:>5} GOs:   {F}\n'.format(N=len(gos), F=fout_allgos))
+        if log is not None:
+            log.write('    taxid {TAXID:>5}\n'.format(TAXID=taxid))
+            log.write('    FOUND {N:>5} GOs:   {F}\n'.format(
+                N=len(gos_all), F=fout_allgos))
     # Get Entrez GeneIDs for cell cycle GOs
     geneids = srch.get_items(gos_all)
-    prt_genes("cell_cycle_genes.txt", geneids, srch, log)
     return geneids
 
-def prt_genes(fout_genes, geneids, srch, log):
-    """Print genes."""
-    from genes_NCBI_hsa_All import GeneID2nt
-    # Print genes related to cell cycle
-    fmtstr = "{GeneID:>9} {Symbol:>16} {description}\n"
-    nts = [GeneID2nt[geneid] for geneid in sorted(geneids) if geneid in GeneID2nt]
-    with open(fout_genes, 'w') as prt:
-        prt_txt(prt, nts, fmtstr)
-        log.write("    WROTE {N:>5} genes: {FOUT}\n".format(FOUT=fout_genes, N=len(nts)))
+def prt_genes(fout_genes, geneids, taxid, log):
+    """Print 'cell cycle' geneids, with or without Symbol and description information."""
+    fin_symbols = "genes_NCBI_{TAXID}_All.py".format(TAXID=taxid)
+    # If gene Symbol information is available, print geneid and Symbol
+    if os.path.isfile(fin_symbols):
+        import importlib
+        module = importlib.import_module(fin_symbols[:-3])
+        GeneID2nt = module.GeneID2nt
+        fmtstr = "{GeneID:>9} {Symbol:<16} {description}\n"
+        nts = [GeneID2nt[geneid] for geneid in sorted(geneids) if geneid in GeneID2nt]
+        with open(fout_genes, 'w') as prt:
+            prt_txt(prt, nts, fmtstr)
+            if log is not None:
+                log.write("    WROTE {N:>5} genes: {FOUT}\n".format(FOUT=fout_genes, N=len(nts)))
+    # Just print geneids
+    else:
+        with open(fout_genes, 'w') as prt:
+            for geneid in geneids:
+                prt.write("{geneid}\n".format(geneid=geneid))
+            if log is not None:
+                log.write("    WROTE {N:>5} genes: {FOUT}\n".format(FOUT=fout_genes, N=len(geneids)))
 
 if __name__ == '__main__':
-    test_cell_cycle()
+    test_cell_cycle(9606)
+    test_cell_cycle(10090)
 
 # Copyright (C) 2010-2016, DV Klopfenstein, H Tang, All rights reserved.
